debugging tests

+++ test/3000_verify-lhash1-push.cpp
@@ -50,7 +50,7 @@ int main() {
     // Verify row position incremented to 1
     size_t position = matrix.debug_get_row_position(row);
 
-    tester.assertEqual(position,1std::format("Row position for row {} is {} expected 1",row,position));
+    tester.assertEqual(position,size_t(1),std::format("Row position for row {} is {} expected 1",row,position));
 
     tester.pass("verify-lhash1-push completed successfully.");
     return EXIT_SUCCESS;
